+++ src/utils/chatStyle.ts
@@ -149,6 +149,15 @@ export const StyleImage = styled.div`
   justify-content: center;
   align-items: center;
 `;
+export const IconImage = styled.div`
+  width: 40px; 
+  height: 40px; 
+  border-radius: 50%;
+  overflow: hidden;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
 
 export const MessageContent = styled(Box) <StyledProps>`
   display: flex;
@@ -166,7 +175,7 @@ export const MessageContent = styled(Box) <StyledProps>`
 export const StyledImage = styled.img`
   width: 100%;
   height: 100%;
-  object-fit: cover;
+  object-fit: scale-down;
 `;
 
 export const StyledAvatar = styled(Avatar) <StyledProps>`